+++ core/model/src/main/java/com/atelbay/money_manager/core/model/Account.kt
@@ -4,6 +4,6 @@ data class Account(
     val id: Long = 0,
     val name: String,
     val currency: String,
-    val balance: Double,
+    val balance: Long,
     val createdAt: Long = 0,
 )
